refactor(app): extract MapInteractionController from main.ts (#787 phase 8d)
Moves handleHexTap (the switch-based executor over resolveMapTapIntent from
phase 8b), handleHexLongPress, and the territory-inspection-panel lifecycle
(openTerritoryInspectionPanel/closeTerritoryInspectionPanel) out of main.ts
into src/app/controllers/map-interaction-controller.ts, verbatim -- this
phase relocates the dispatcher, it does not change precedence (owned
entirely by resolveMapTapIntent, phase 8a) or add new behavior.

Verified via an exhaustive line-by-line diff of the extracted body against
the pre-8d main.ts source (the same technique that caught 2 real bugs
during phase 8c's review): only one line differed, and it was a false
positive from the normalization script's inability to express object-literal
shorthand syntax via regex -- the actual code was already correct. SFX,
bus.emit, session.commit/setStateWithoutRefresh, showNotification, and
updateHUD call counts all match the original exactly.

getElementById substitutes for eight distinct document.getElementById(...)
calls this code used to make directly (info-panel, tech-panel, city-panel,
espionage-panel, diplomacy-panel, marketplace-panel, council-panel,
territory-inspection-panel) -- one generic getter matching the native
signature, not eight single-purpose named getters, since Phase 11's
port-purity test bans document.getElementById in src/app/controllers/*.

bus is typed as the concrete EventBus class, not a narrowed Pick<'emit'>,
matching the lesson from phase 8c's review: two downstream calls
(beginConfirmedForeignCityEntry, emitMinorCivQuestTransitions) require it,
and EventBus has a private field so no object literal can structurally
satisfy it either way.

Proactively fixed two pieces of now-stale documentation while moving the
code: a docblock comment and a thrown-error message both referenced
`openTerritoryInspectionPanel(coord)` as a directly-callable main.ts
function, which it no longer is (it's now a private closure inside the
controller, reachable only through handleHexLongPress). Also removed a
latent dead import (renderSelectedUnitInfo) that phase 8c had already
orphaned -- only a prose comment referenced it, which moved with the code
in this phase, fully exposing the dead import.

tests/main.integration.test.ts: two grep assertions broke as an unavoidable
consequence of the move (both anchored on `function handleHexTap(` /
`case '...': {` text that no longer exists in main.ts). Retargeted both to
read src/app/controllers/map-interaction-controller.ts instead, and added a
"map interaction controller wiring" wiring-sanity test (mirroring phase
8c's pattern) proving main.ts constructs the controller and routes
touch/mouse input through it, with the four moved function declarations
confirmed absent from main.ts.

Scope note: the plan's Phase 8d Step 5 also lists "player combat wiring"
(3), "shared city founding wiring" (1), "shared unit upgrade wiring" (1),
and "shared city assault wiring" (4) as grep assertions to convert to real
tests. Left those 9 as-is -- they were not broken by this change (their
target functions, executeAttack/foundCityAction/executeUpgrade/
beginPlayerCityAssault/executeMinorCivConquest, all remain main.ts-local
dependencies passed into MapInteractionController, not files this phase
creates or moves). Converting them to genuine behavioral tests would mean
extracting six more main.ts functions into testable modules, which is
outside MapInteractionController's actual file scope and would be its own
follow-up if wanted.

32 imports that existed only to serve the moved code were removed from
main.ts; verified against the pre-8d baseline that no new dead imports
were introduced and the 13 pre-existing ones are untouched.

Verification: yarn build, yarn test (470 files / 7625 tests), and
yarn test:web-smoke (8/8, including the water-recovery e2e spec that
directly exercises the moved blocked-tap dispatch) all green. Full-dimension
inline review performed before opening this PR; no issues found beyond the
housekeeping already folded into this commit.
